Move stories to the same directory as components.
