+++ frontend/src/components/Button/Button.stories.tsx
@@ -3,8 +3,8 @@ import { text } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react-native';
 import React from 'react';
 import { Text } from 'react-native';
-import Button from '../components/Button';
-import CenterView from '../components/CenterView';
+import Button from '.';
+import CenterView from '../CenterView';
 
 storiesOf('Button', module)
   .addDecorator((getStory) => <CenterView>{getStory()}</CenterView>)
