+++ frontend/src/components/Welcome/Welcome.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { linkTo } from '@storybook/addon-links';
 import { storiesOf } from '@storybook/react-native';
-import Welcome from './Welcome';
+import Welcome from '.';
 
 storiesOf('Welcome', module).add('to Storybook', () => <Welcome showApp={linkTo('Button')} />);
