+++ frontend/src/components/stories/index.ts
@@ -1,0 +1,2 @@
+import '../Welcome/Welcome.stories';
+import '../Button/Button.stories';
