--- frontend/src/stories/index.ts
@@ -1,2 +1,0 @@
-import './Welcome.stories';
-import './Button.stories';
